master: Implement template of Menus, Begin implement finish feature for create Expense for nutrition

+++ Kindergarten/Kindergarten/Models/Entities/Dish.cs
@@ -18,10 +18,10 @@ namespace Kindergarten.Models.Entities
         public string Name { get; set; }
 
         [Required]
-        public int DishNurseryNorm { get; set; }
+        public float DishNurseryNorm { get; set; }
 
         [Required]
-        public int DishYardNorm { get; set; }
+        public float DishYardNorm { get; set; }
 
         public ICollection<DishItem> DishItems { get; set; }
